+++ app/src/main/java/app/revanced/manager/ui/component/InstallerStatusDialog.kt
@@ -2,26 +2,17 @@ package app.revanced.manager.ui.component
 
 import android.annotation.SuppressLint
 import android.content.pm.PackageInstaller
-import androidx.annotation.RequiresApi
 import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.ErrorOutline
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
-import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
 import app.revanced.manager.R
 import app.revanced.manager.ui.model.InstallerModel
 import com.github.materiiapps.enumutil.FromValue
@@ -43,9 +34,6 @@ fun InstallerStatusDialog(installerStatus: Int, model: InstallerModel, onDismiss
         dismissButton = {
             dialogKind.dismissButton?.invoke(model, onDismiss)
         },
-        icon = {
-            Icon(dialogKind.icon, null)
-        },
         title = {
             Text(
                 text = stringResource(dialogKind.title),
@@ -54,12 +42,7 @@ fun InstallerStatusDialog(installerStatus: Int, model: InstallerModel, onDismiss
             )
         },
         text = {
-            Column(
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.spacedBy(16.dp),
-            ) {
-                Text(stringResource(dialogKind.contentStringResId))
-            }
+            Text(stringResource(dialogKind.contentStringResId))
         }
     )
 }
@@ -85,7 +68,6 @@ enum class DialogKind(
     val flag: Int,
     val title: Int,
     @param:StringRes val contentStringResId: Int,
-    val icon: ImageVector = Icons.Outlined.ErrorOutline,
     val confirmButton: InstallerStatusDialogButton = installerStatusDialogButton(R.string.ok),
     val dismissButton: InstallerStatusDialogButton? = null,
 ) {
@@ -93,58 +75,47 @@ enum class DialogKind(
         flag = PackageInstaller.STATUS_FAILURE,
         title = R.string.installation_failed_dialog_title,
         contentStringResId = R.string.installation_failed_description,
-        confirmButton = installerStatusDialogButton(R.string.install_app) { model ->
-            model.install()
-        }
-    ),
-    FAILURE_ABORTED(
-        flag = PackageInstaller.STATUS_FAILURE_ABORTED,
-        title = R.string.installation_cancelled_dialog_title,
-        contentStringResId = R.string.installation_aborted_description,
-        confirmButton = installerStatusDialogButton(R.string.install_app) { model ->
-            model.install()
-        }
+        confirmButton = installerStatusDialogButton(R.string.try_again) { it.install() },
+        dismissButton = installerStatusDialogButton(R.string.cancel),
     ),
     FAILURE_BLOCKED(
         flag = PackageInstaller.STATUS_FAILURE_BLOCKED,
         title = R.string.installation_blocked_dialog_title,
         contentStringResId = R.string.installation_blocked_description,
+        dismissButton = installerStatusDialogButton(R.string.cancel),
     ),
     FAILURE_CONFLICT(
         flag = PackageInstaller.STATUS_FAILURE_CONFLICT,
         title = R.string.installation_conflict_dialog_title,
         contentStringResId = R.string.installation_conflict_description,
-        confirmButton = installerStatusDialogButton(R.string.reinstall) { model ->
-            model.reinstall()
-        },
+        confirmButton = installerStatusDialogButton(R.string.reinstall) { it.reinstall() },
         dismissButton = installerStatusDialogButton(R.string.cancel),
     ),
     FAILURE_INCOMPATIBLE(
         flag = PackageInstaller.STATUS_FAILURE_INCOMPATIBLE,
         title = R.string.installation_incompatible_dialog_title,
         contentStringResId = R.string.installation_incompatible_description,
+        dismissButton = installerStatusDialogButton(R.string.ok),
     ),
     FAILURE_INVALID(
         flag = PackageInstaller.STATUS_FAILURE_INVALID,
         title = R.string.installation_invalid_dialog_title,
         contentStringResId = R.string.installation_invalid_description,
-        confirmButton = installerStatusDialogButton(R.string.reinstall) { model ->
-            model.reinstall()
-        },
+        confirmButton = installerStatusDialogButton(R.string.reinstall) { it.reinstall() },
         dismissButton = installerStatusDialogButton(R.string.cancel),
     ),
     FAILURE_STORAGE(
         flag = PackageInstaller.STATUS_FAILURE_STORAGE,
         title = R.string.installation_storage_issue_dialog_title,
         contentStringResId = R.string.installation_storage_issue_description,
+        dismissButton = installerStatusDialogButton(R.string.ok),
     ),
     FAILURE_TIMEOUT(
         flag = @SuppressLint("InlinedApi") PackageInstaller.STATUS_FAILURE_TIMEOUT,
         title = R.string.installation_timeout_dialog_title,
         contentStringResId = R.string.installation_timeout_description,
-        confirmButton = installerStatusDialogButton(R.string.install_app) { model ->
-            model.install()
-        },
+        confirmButton = installerStatusDialogButton(R.string.try_again) { it.install() },
+        dismissButton = installerStatusDialogButton(R.string.cancel),
     );
 
     // Needed due to the @FromValue annotation.